more compilation warnings fixed

+++ opengl_acquisition_shared_library/opengl_depth_and_color_renderer/src/Library/Rendering/ShaderPipeline/computeShader.h
@@ -12,7 +12,7 @@ struct computeShaderObject
   int computeShaderProgram;
 
   char * compMem;
-  unsigned long compMemLength;
+  int compMemLength;
 };
